@@ -2,25 +2,21 @@ source 'https://rubygems.org'
 
 ruby '2.0.0'
 
-gem 'haml-rails'
+gem 'angularjs-rails'
+gem 'bourbon'
+gem 'coffee-rails'
 gem 'delayed_job_active_record'
+gem 'haml-rails'
 gem 'jquery-rails'
 gem 'octokit'
 gem 'omniauth-github'
 gem 'pg'
-gem 'rails', '3.2.16'
+gem 'rails', '4.0.0'
 gem 'rollbar'
 gem 'rubocop'
-gem 'strong_parameters'
+gem 'sass-rails'
+gem 'uglifier', '>= 1.0.3'
 gem 'unicorn'
-
-group :assets do
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'sass-rails', '~> 3.2.3'
-  gem 'uglifier', '>= 1.0.3'
-  gem 'bourbon'
-  gem 'angularjs-rails'
-end
 
 group :development, :test do
   gem 'debugger'
@@ -34,4 +30,8 @@ group :test do
   gem 'factory_girl_rails'
   gem 'shoulda-matchers'
   gem 'webmock'
+end
+
+group :staging, :production do
+  gem 'rails_12factor'
 end
