@@ -9,22 +9,25 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130621203400) do
+ActiveRecord::Schema.define(version: 20131212192202) do
 
-  create_table "builds", :force => true do |t|
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "builds", force: true do |t|
     t.text     "violations"
     t.integer  "repo_id"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  add_index "builds", ["repo_id"], :name => "index_builds_on_repo_id"
+  add_index "builds", ["repo_id"], name: "index_builds_on_repo_id", using: :btree
 
-  create_table "delayed_jobs", :force => true do |t|
-    t.integer  "priority",   :default => 0
-    t.integer  "attempts",   :default => 0
+  create_table "delayed_jobs", force: true do |t|
+    t.integer  "priority",   default: 0
+    t.integer  "attempts",   default: 0
     t.text     "handler"
     t.text     "last_error"
     t.datetime "run_at"
@@ -32,28 +35,32 @@ ActiveRecord::Schema.define(:version => 20130621203400) do
     t.datetime "failed_at"
     t.string   "locked_by"
     t.string   "queue"
-    t.datetime "created_at",                :null => false
-    t.datetime "updated_at",                :null => false
+    t.datetime "created_at",             null: false
+    t.datetime "updated_at",             null: false
   end
 
-  add_index "delayed_jobs", ["priority", "run_at"], :name => "delayed_jobs_priority"
+  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
 
-  create_table "repos", :force => true do |t|
-    t.integer "github_id",                           :null => false
-    t.boolean "active",           :default => false, :null => false
-    t.integer "user_id",                             :null => false
+  create_table "memberships", force: true do |t|
+    t.integer "user_id", null: false
+    t.integer "repo_id", null: false
+  end
+
+  add_index "memberships", ["repo_id", "user_id"], name: "index_memberships_on_repo_id_and_user_id", using: :btree
+
+  create_table "repos", force: true do |t|
+    t.integer "github_id",                        null: false
+    t.boolean "active",           default: false, null: false
     t.integer "hook_id"
-    t.string  "name",                                :null => false
-    t.string  "full_github_name",                    :null => false
+    t.string  "name",                             null: false
+    t.string  "full_github_name",                 null: false
   end
 
-  add_index "repos", ["user_id", "github_id"], :name => "index_repos_on_user_id_and_github_id", :unique => true
-
-  create_table "users", :force => true do |t|
-    t.datetime "created_at",      :null => false
-    t.datetime "updated_at",      :null => false
-    t.string   "github_username", :null => false
-    t.string   "remember_token",  :null => false
+  create_table "users", force: true do |t|
+    t.datetime "created_at",      null: false
+    t.datetime "updated_at",      null: false
+    t.string   "github_username", null: false
+    t.string   "remember_token",  null: false
     t.string   "github_token"
   end
 
