@@ -21,10 +21,14 @@ class ApplicationController < ActionController::Base
   end
 
   def set_csrf_cookie_for_ng
-    cookies['XSRF-TOKEN'] = form_authenticity_token if protect_against_forgery?
+    if protect_against_forgery?
+      cookies['XSRF-TOKEN'] = form_authenticity_token
+    end
   end
 
+  protected
+
   def verified_request?
-    super || form_authenticity_token == request.headers['X_XSRF_TOKEN']
+    super || form_authenticity_token == request.headers['X-XSRF-TOKEN']
   end
 end
