@@ -1,16 +1,15 @@
 class Repo < ActiveRecord::Base
   include ActiveModel::ForbiddenAttributesProtection
 
-  attr_accessible :active, :full_github_name, :github_id, :hook_id, :name
-
-  belongs_to :user
+  has_many :memberships
+  has_many :users, through: :memberships
   has_many :builds
 
   validates :name, presence: true
   validates :full_github_name, presence: true
-  validates :github_id, uniqueness: { scope: :user_id }, presence: true
+  validates :github_id, uniqueness: true, presence: true
 
-  scope :active, where(active: true)
+  scope :active, -> { where(active: true) }
 
   def activate
     update_attribute(:active, true)
@@ -21,6 +20,6 @@ class Repo < ActiveRecord::Base
   end
 
   def github_token
-    user.github_token
+    users.first.github_token
   end
 end
