@@ -6,10 +6,11 @@ describe ReposController, '#update' do
       activator = double(:repo_activator, activate: true)
       RepoActivator.stub(new: activator)
       user = create(:user)
-      repo = create(:repo, user: user)
+      repo = create(:repo)
+      create(:membership, user: user, repo: repo)
       stub_sign_in(user)
 
-      put(:update, id: repo.id, active: true)
+      patch(:update, id: repo.id, active: true, format: :json)
 
       expect(activator).to have_received(:activate).with(repo)
     end
@@ -20,10 +21,11 @@ describe ReposController, '#update' do
       activator = double(:repo_activator, deactivate: true)
       RepoActivator.stub(new: activator)
       user = create(:user)
-      repo = create(:repo, user: user)
+      repo = create(:repo)
+      create(:membership, user: user, repo: repo)
       stub_sign_in(user)
 
-      put(:update, id: repo.id, active: false)
+      patch(:update, id: repo.id, active: false, format: :json)
 
       expect(activator).to have_received(:deactivate).with(repo)
     end
